requirements.txt
asgiref==3.7.2
contourpy==1.1.1
cycler==0.12.1
Django==4.2.5
fonttools==4.43.1
imageio==2.31.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.1
networkx==3.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0.post7
sqlparse==0.4.4
threadpoolctl==3.2.0
tifffile==2023.8.30
tzdata==2023.3
